fix: varredura defensiva — timeout global, relatório e score float
1. app.py: capturar TimeoutError do iterador as_completed() — antes,
   timeout global crashava o loop sem finalizar resultados. Agora
   preenche perguntas pendentes com erro e continua
2. app.py: try/except na geração e download do relatório Excel —
   falha de I/O não derruba mais a seção de resultados
3. report_handler.py: int() → float() na formatação condicional —
   previne ValueError com scores decimais vindos como string

+++ report_handler.py
@@ -76,7 +76,7 @@ def generate_report(results: list[dict], rag_metadata: dict | None = None, score
             score_val = score_cell.value
             if score_val is not None and score_val != "":
                 try:
-                    score_num = int(score_val)
+                    score_num = float(score_val)
                 except (ValueError, TypeError):
                     continue
                 if score_num >= 70:
